Correct handling of value type A in F[A]
A drawing is some type F[A] for which there is an image algebras of type F. The
type A is the value the image produces when evaluated. It has the effect
of also drawing the image.

Previously the type A was scope to the image algebra traits, not to the algebra
methods. This basically meant each algebra could only handle images with a
fixed value type, which is bogus. Shifting the type A to methods makes
everything work that way it should. This change does this.