Create finer grained abstractions for the rendering process
Add an `Engine` abstraction that describes how to create some kind of value from
a `Frame` (which describes an area we want to render to), and then allows
rendering to this value. Previously there was no abstraction for these
steps---they were rolled into one in the `Renderer` abstraction.

Exposing these abstractions allows greater control over rendering, with an
ultimate goal of enabling animations.