Restructure animation framework
Animation is much simpler using Monix, but I didn't want to add a Monix
dependency to the main project so I've split animation into a separate project.

Rename Animator -> Animation

Shift the code that does the actual animation into the Animation typeclass, so
this is completely encapsulated. Animation syntax now just delegates to the
typeclass instance.