Initial implementation of Image algebraic data type
The `Image` ADT mirrors the old Doodle language (currently with fewer features
as this version of Doodle doesn't yet support all the features of the previous
version). This language is not extensible like the tagless final algebra
language is, but it much more convenient to use for those cases where it applies
as there is no need for the type gymnastics the tagless final algebra needs.
This is particularly relevant when using Doodle for teaching.